+++ src/backend/middlewares/erklaering-interceptor.test.ts
@@ -4,7 +4,7 @@ import { DeepPartial } from 'ts-essentials';
 
 import { LocaleType } from '@navikt/familie-sprakvelger';
 
-import { ISøknadKontrakt } from '../../frontend/typer/kontrakt/generelle';
+import { ISøknadKontrakt } from '../../frontend/typer/kontrakt/v6';
 import { erklaeringInterceptor, hentSpråkteksterAlleSpråk } from './erklaering-interceptor';
 
 describe('erklaering-interceptor', () => {
