+++ src/frontend/components/SøknadsSteg/Dokumentasjon/filopplaster/useFilopplaster.tsx
@@ -6,7 +6,7 @@ import { fromBlob } from 'file-type/browser';
 import { useLastRessurserContext } from '../../../../context/LastRessurserContext';
 import Miljø from '../../../../Miljø';
 import { IDokumentasjon, IVedlegg } from '../../../../typer/dokumentasjon';
-import { Dokumentasjonsbehov } from '../../../../typer/kontrakt/generelle';
+import { Dokumentasjonsbehov } from '../../../../typer/kontrakt/dokumentasjon';
 import { formaterFilstørrelse } from '../../../../utils/dokumentasjon';
 import SpråkTekst from '../../../Felleskomponenter/SpråkTekst/SpråkTekst';
 import { konverter } from './konverteringService';
