+++ src/frontend/hooks/useSendInnSkjema.test.tsx
@@ -1,7 +1,8 @@
 import { renderHook } from '@testing-library/react-hooks';
 import { act } from 'react-dom/test-utils';
 
-import { ESivilstand, ISøknadKontrakt } from '../typer/kontrakt/generelle';
+import { ESivilstand } from '../typer/kontrakt/generelle';
+import { ISøknadKontrakt } from '../typer/kontrakt/v6';
 import { hentSivilstatusSpråkId } from '../utils/språk';
 import {
     mekkGyldigUtvidetSøknad,
