+++ src/frontend/typer/kontrakt/søker.ts
@@ -1,0 +1,21 @@
+import {
+    ESivilstand,
+    IAdresse,
+    IKontraktNåværendeSamboer,
+    IKontraktTidligereSamboer,
+    ISøknadsfelt,
+    IUtenlandsperiodeIKontraktFormat,
+    SpørsmålMap,
+} from './generelle';
+
+export interface ISøknadKontraktSøker {
+    ident: ISøknadsfelt<string>;
+    navn: ISøknadsfelt<string>;
+    statsborgerskap: ISøknadsfelt<string[]>;
+    adresse: ISøknadsfelt<IAdresse>;
+    sivilstand: ISøknadsfelt<ESivilstand>;
+    spørsmål: SpørsmålMap;
+    tidligereSamboere: ISøknadsfelt<IKontraktTidligereSamboer>[];
+    nåværendeSamboer: ISøknadsfelt<IKontraktNåværendeSamboer> | null;
+    utenlandsperioder: ISøknadsfelt<IUtenlandsperiodeIKontraktFormat>[];
+}
