+++ src/frontend/typer/søknad.ts
@@ -7,7 +7,8 @@ import { genererInitiellDokumentasjon } from '../utils/dokumentasjon';
 import { IBarnMedISøknad } from './barn';
 import { INøkkelPar } from './common';
 import { dokumentasjonsbehovTilSpråkId, IDokumentasjon } from './dokumentasjon';
-import { Dokumentasjonsbehov, ESivilstand, ESøknadstype } from './kontrakt/generelle';
+import { Dokumentasjonsbehov } from './kontrakt/dokumentasjon';
+import { ESivilstand, ESøknadstype } from './kontrakt/generelle';
 import { IBarn, ISøker } from './person';
 import { ISøknadSpørsmål } from './spørsmål';
 import { Årsak } from './utvidet';
